refactor: split streaming sse helpers
