+++ presets/ragengine/streaming/upstream.py
@@ -1,4 +1,3 @@
-import json
 from typing import Any
 
 SSE_DATA_PREFIX = "data:"
@@ -27,23 +26,3 @@ def extract_delta_content(payload: dict[str, Any]) -> str | None:
 
     content = delta.get("content")
     return content if isinstance(content, str) else None
-
-
-def build_sse_data_chunk(payload: dict[str, Any]) -> str:
-    return f"{SSE_DATA_PREFIX} {json.dumps(payload, separators=(',', ':'))}\n\n"
-
-
-def build_sse_done_chunk() -> str:
-    return f"{SSE_DATA_PREFIX} {SSE_DONE_MARKER}\n\n"
-
-
-def build_block_sse_chunk(message: str) -> str:
-    payload = {
-        "choices": [
-            {
-                "delta": {"content": message},
-                "finish_reason": "content_filter",
-            }
-        ]
-    }
-    return build_sse_data_chunk(payload)
